Better collision

+++ src/World/World.cpp
@@ -5,7 +5,7 @@
 #include "World.h"
 
 namespace World {
-    World::World(int width, int height)
+    World::World(float width, float height)
     : m_engine{width, height}, m_spawner{{width - 40.0f, 20.0f}, 100, 20.0f, 0.1f} {
         m_entities.reserve(50);
     }
